break up default step for errorbar in separate file
